backend Novo/classigarapava_backend_render/requirements.txt
Flask==3.1.1
Flask-CORS==6.0.1
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.41
PyJWT==2.8.0
bcrypt==4.2.1
python-dotenv==1.0.1
gunicorn==23.0.0
Werkzeug==3.1.3

